+++ engine/src/main/java/org/terasology/math/SpiralIterable.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.math.iterator;
+package org.terasology.math;
 
 import com.google.common.base.Preconditions;
 import org.joml.Vector2i;
